refactor: extract condition disabled states
