+++ extension/src/entrypoints/popup/pages/profiles/components/Header/components/AddRuleOptionDialog/conditions.ts
@@ -1,68 +1,20 @@
-import type { AddRuleOptionDialogItem, AddRuleOptionDialogTab } from "./shared";
+import type { AddRuleOptionDialogTab } from "./shared";
 import { uuidv7 } from "uuidv7";
 import { useI18n } from "vue-i18n";
 import { useProfilesStore } from "@/entrypoints/popup/stores/useProfilesStore";
 import { getCurrentTabHost, getCurrentTabHostname } from "@/lib/currentTab";
 import { getDefaultFilterValueByHost } from "@/lib/filter";
-import {
-  getEnabledState,
-  withDisabledState,
-} from "./shared";
+import { withDisabledState } from "./shared";
+import { useConditionDisabledStates } from "./useConditionDisabledStates";
 
 export function useCreateConditionTab(): AddRuleOptionDialogTab {
   const { t } = useI18n();
   const profilesStore = useProfilesStore();
-
-  function getConditionAlreadyAddedDisabledState(conditionExists: boolean) {
-    if (!conditionExists) {
-      return getEnabledState();
-    }
-
-    return {
-      disabled: true,
-      disabledTooltip: t("addRuleOptionDialog.disabledTooltip.conditionAlreadyAdded"),
-    };
-  }
-
-  function withConditionAlreadyAddedDisabledState<T extends AddRuleOptionDialogItem>(
-    item: T,
-    conditionExists: () => boolean,
-  ) {
-    return withDisabledState(
-      item,
-      () => getConditionAlreadyAddedDisabledState(conditionExists()),
-    );
-  }
-
-  function getUrlFilterDisabledState() {
-    if (profilesStore.selectedProfile.filters.urlFilter?.length) {
-      return getConditionAlreadyAddedDisabledState(true);
-    }
-
-    if (profilesStore.selectedProfile.filters.regexFilter?.length) {
-      return {
-        disabled: true,
-        disabledTooltip: t("addRuleOptionDialog.disabledTooltip.urlFilterBlockedByRegexFilter"),
-      };
-    }
-
-    return getEnabledState();
-  }
-
-  function getRegexFilterDisabledState() {
-    if (profilesStore.selectedProfile.filters.regexFilter?.length) {
-      return getConditionAlreadyAddedDisabledState(true);
-    }
-
-    if (profilesStore.selectedProfile.filters.urlFilter?.length) {
-      return {
-        disabled: true,
-        disabledTooltip: t("addRuleOptionDialog.disabledTooltip.regexFilterBlockedByUrlFilter"),
-      };
-    }
-
-    return getEnabledState();
-  }
+  const {
+    getRegexFilterDisabledState,
+    getUrlFilterDisabledState,
+    withConditionAlreadyAddedDisabledState,
+  } = useConditionDisabledStates();
 
   return {
     label: t("addRuleOptionDialog.tabs.conditions"),
